Use different (Dagger) modules instead of the enum to choose dependency implementation.
Seems more general and clean to me. Simpler to apply when all the other dependencies are in place. The "magic" enum is no longer necessary, the dependency graph is defined in Dagger only. The (gradle) modules can "export" (make public) the (Dagger) modules the clients (the apps) are supposed to use.

This method also changes how each (gradle) module uses the dependencies. `feat2` no longer exports `dep`, which means that the app has to depend on some concrete implementation or on the module itself, depending on what's better. And we avoid the transitive dependency.