Use constructor injection on fragments
Used a FragmentFactory to allow Dagger to inject the constructor of fragments (see InjectFragmentFactory.kt).

This needs a fair amount of boiler plate code to make it work. Some of it can be done only once: the factory setup and all that is in the `dicore` (gradle) module. It does force the app modules to inherit from the base classes there. But we also need an extra module for each fragment/feature that will be injected so it can be added to the factory (see Feature2BindsModule.kt). If we want a different scope for the fragment/feature, the factory will be responsible for creating the corresponding Dagger (sub) component.

One important thing to note is that dependency injection on the fragment should be used **very** sparingly, mainly to inject/provide the view model(s), that is where dependency injection will actually be more heavily used. So, as Uncle Ben said, "with great power, comes great responsibility": this is not to be abused!

On the plus side, the interface that the application needed to implement for each feature (ex: Feature2ComponentProvider) is gone and the inject method on the fragment is no longer necessary, replaced by a method to provide/create one.

I created a couple of "local" dependencies on `feat2` to demonstrate/test the use of the scope. It was mainly useful for me to make sure ai understood things, but I left them here as it might help others.

I also moved the base dagger version variable to the base gradle file, to centralize the dependency.

Some references:

https://proandroiddev.com/inject-into-android-component-constructor-4f5ddd27d06
https://proandroiddev.com/constructor-injection-into-fragments-with-respect-of-scopes-311c0258d678
https://dagger.dev/dev-guide/multibindings